refactor firehose message treatment
