+++ cmd/rabtap/message_test.go
@@ -1,0 +1,5 @@
+// Copyright (C) 2017 Jan Delgado
+
+package main
+
+// TODO
