+++ cmd/rabtap/message_transformer.go
@@ -1,16 +1,16 @@
-// transform RabtapPersistentMessages
+// compose message providers
 package main
 
-// MessageTransformer transforms a RabtapPersistentMessage
-type MessageTransformer func(RabtapPersistentMessage) (RabtapPersistentMessage, error)
+// MessageTransformer transforms the given message
+type MessageTransformer func(m RabtapPersistentMessage) (RabtapPersistentMessage, error)
 
-// NewComposingMessageTransformer returns a new message transformer that computes
-// m' = f(g(m))
-func NewComposingMessageTransformer(f, g MessageTransformer) MessageTransformer {
-	return func(m RabtapPersistentMessage) (RabtapPersistentMessage, error) {
-		m, err := g(m)
+// NewTransformingMessageProvider returns a new message provider that computes
+// m = t(f()), i.e. that applies the transformer to the message provided by f.
+func NewTransformingMessageProvider(t MessageTransformer, f MessageProviderFunc) MessageProviderFunc {
+	return func() (RabtapPersistentMessage, error) {
+		m, err := f()
 		if err == nil {
-			return f(m)
+			return t(m)
 		}
 		return RabtapPersistentMessage{}, err
 	}
